content/canvas/test/test_2d.imageData.get.order.cols.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.get.order.cols</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#fff';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#000';
ctx.fillRect(0, 0, 2, 50);
var imgdata = ctx.getImageData(0, 0, 10, 10);
ok(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");
ok(imgdata.data[Math.floor(imgdata.width/2*4)] === 255, "imgdata.data[Math.floor(imgdata.width/2*4)] === 255");
ok(imgdata.data[(imgdata.height/2)*imgdata.width*4] === 0, "imgdata.data[(imgdata.height/2)*imgdata.width*4] === 0");

if (!_deferred) SimpleTest.finish();
});
</script>
